Refactor llm_embedding: modularize and clean up codebase
- Extract configuration to config.py
- Move data handling to data.py (datasets, data loading)
- Separate loss functions to loss.py (InfoNCE loss)
- Create utils.py for metrics, logging, and evaluation utilities
- Simplify training.py to focus on core training logic
- Streamline inference.py for cleaner inference workflow
- Reduce verbosity by ~400 lines while maintaining functionality
- Improve separation of concerns and code maintainability
